requirements.txt
flask == 2.3.0
spotipy == 2.23.0
request == 2.32.3